refactor: Update paths for layout imports
