+++ src/components/Checkbox/Checkbox.tsx
@@ -10,7 +10,7 @@ import type {
 
 import { cn } from '@utils';
 
-import { Grid } from '@layouts';
+import { Grid } from '@layouts/Grid';
 
 import { Text } from '../Text';
