+++ src/components/Switch/Switch.tsx
@@ -8,7 +8,7 @@ import type {
 
 import { cn } from '@utils';
 
-import { Flex } from '@layouts';
+import { Flex } from '@layouts/Flex';
 
 export type SwitchProps = StrictOmit<
   ComponentPropsWithRef<'button'>,
